backend/requirements.txt
aiostream
annotated-types
anyio
beanie
certifi
click
dnspython
email_validator
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
lazy-model
markdown-it-py
MarkupSafe
mdurl
motor
orjson
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
pymongo
python-dotenv
python-multipart
PyYAML
rich
rich-toolkit
shellingham
sniffio
starlette
toml
typer
typing-inspection
typing_extensions
ujson
uvicorn
uvloop
watchfiles
websockets
bcrypt
pydantic[email]
selenium
webdriver-manager
beautifulsoup4
lxml
vaderSentiment